Create SleepingBarbarProblem
The Sleeping-Barber Problem. A barbershop consists of a waiting room with n chairs
and a barber room with one barber chair. If there are no customers to be served the barber
goes to sleep. If a customer enters the barbershop and all chairs are occupied then the
customer leaves the shop. If the barber is busy but chairs are available, then the customer sits
in one of the free chairs. If the barber is asleep, the customer wakes up the barber. Write a
program to coordinate the barber and the customers.